+++ src/routes/v2/projects/[projectId]/datasets/[datasetId]/+page.svelte
@@ -48,6 +48,76 @@
 	/** @type {CreateUpdateImageModal|undefined} */
 	let imageModal = undefined;
 
+	$: plates = Array.isArray(imagePage.attributes['plate'])
+		? imagePage.attributes['plate'].sort()
+		: [];
+	let selectedPlate = '';
+	let platePath = '';
+	let platePathLoading = false;
+	let platePathError = '';
+
+	$: if (plates.length > 0 && selectedPlate !== '') {
+		computePlatePath();
+	} else {
+		platePath = '';
+		platePathError = '';
+	}
+
+	async function computePlatePath() {
+		platePathError = '';
+		if (!plates.includes(selectedPlate)) {
+			selectedPlate = '';
+			return;
+		}
+		let imageWithPlate = imagePage.images.find((i) => i.attributes['plate'] === selectedPlate);
+		if (!imageWithPlate) {
+			platePathLoading = true;
+			imageWithPlate = await loadImageForSelectedPlate();
+			platePathLoading = false;
+		}
+		if (imageWithPlate) {
+			// Removes the last 3 elements from the path
+			platePath = imageWithPlate.zarr_url.split('/').slice(0, -3).join('/');
+			if (!platePath) {
+				platePathError = `Unable to load plate URL from zarr URL ${imageWithPlate.zarr_url}`;
+			}
+		} else {
+			platePath = '';
+			platePathError = 'Unable to load plate URL. No image found for the selected plate.';
+		}
+	}
+
+	/**
+	 * @returns {Promise<import('$lib/types-v2').Image|undefined>}
+	 */
+	async function loadImageForSelectedPlate() {
+		const params = { filters: { attributes: { plate: selectedPlate } } };
+		const headers = new Headers();
+		headers.set('Content-Type', 'application/json');
+		const response = await fetch(
+			`/api/v2/project/${projectId}/dataset/${dataset.id}/images/query?page=1&page_size=1`,
+			{
+				method: 'POST',
+				headers,
+				credentials: 'include',
+				body: JSON.stringify(params)
+			}
+		);
+		if (!response.ok) {
+			console.error(`Unable to load image for plate ${selectedPlate}`);
+			return undefined;
+		}
+		/** @type {import('$lib/types-v2').ImagePage}*/
+		const result = await response.json();
+		if (result.images.length === 0) {
+			console.error(
+				`Unable to load image for plate ${selectedPlate}. Server replied with empty list`
+			);
+			return undefined;
+		}
+		return result.images[0];
+	}
+
 	onMount(() => {
 		loadAttributesSelectors();
 		loadTypesSelector();
@@ -428,6 +498,45 @@
 		<a id="downloadDatasetButton" class="d-none">Download dataset link</a>
 	</div>
 </div>
+
+{#if vizarrViewerUrl && plates.length > 0}
+	<div class="border border-info rounded bg-light p-3 mt-2">
+		<div class="row mb-2">
+			<div class="col">
+				This dataset contains {plates.length}
+				{plates.length === 1 ? 'plate' : 'plates'}. Select which plate you want to view
+			</div>
+		</div>
+		<div class="row row-cols-md-auto g-3 align-items-center">
+			<div class="col-12">
+				<select class="form-select" aria-label="Select plate" bind:value={selectedPlate}>
+					<option value="">Select...</option>
+					{#each plates as plate}
+						<option>{plate}</option>
+					{/each}
+				</select>
+			</div>
+			<div class="col-12">
+				{#if platePath}
+					<a
+						href="{vizarrViewerUrl}?source={vizarrViewerUrl}data{platePath}"
+						class="btn btn-info me-2"
+						target="_blank"
+						class:disabled={platePathLoading}
+					>
+						<i class="bi bi-eye" />
+						View plate
+					</a>
+				{:else if platePathError}
+					<span class="text-danger">{platePathError}</span>
+				{/if}
+				{#if platePathLoading}
+					<span class="spinner-border spinner-border-sm" role="status" aria-hidden="true" />
+				{/if}
+			</div>
+		</div>
+	</div>
+{/if}
 
 {#if !showTable}
 	<p class="fw-bold ms-4 mt-5">No entries in the image list yet</p>